Add date display fx to utils folder
