+++ client/src/views/Dashboard.tsx
@@ -13,6 +13,7 @@ import axios from 'axios';
 import DeleteButton from '../components/buttons/DeleteButton';
 import EventModelForView from '../models/EventModelForView';
 import Organizer from '../models/Organizer';
+import dateDisplay from '../utils/dates/DateDisplay'
 axios.defaults.withCredentials = true;
 
 
@@ -64,27 +65,6 @@ const Dashboard = () => {
             .catch(errors => console.error(errors));
     },[]);
 
-    function dateChange(date: string){
-
-        date = date.slice(0,19);
-        
-        const MM: Array<string> = ["January", "February","March","April","May","June","July","August","September","October","November", "December"]
-        const year: string = date.slice(0, 4);
-        const  month: string = date.slice(5, 7);
-        const day: string = date.slice(8, 10);
-
-        let hour: string = String(parseInt(date.slice(11,13)));
-        if (parseInt(hour) < 10){ hour = "0"+hour }
-
-        let min: string = String(parseInt(date.slice(14, 17)));
-        if (parseInt(min) < 10){ min = "0"+min }
-        
-        const a_p: string = parseInt(hour) < 12 ? 'AM' : 'PM'
-        
-        return ''+MM[parseInt(month)-1]+' '+day+', '+year+' '+hour+':'+min+' '+a_p;
-        
-        }
-
     const removeFromDom = (_id: string) => {
         setAllEvents(allEvents.filter( oneEvent => oneEvent._id !== _id));
     }
@@ -127,7 +107,7 @@ const Dashboard = () => {
 
                 <TableCell align="right">{oneEvent.type}</TableCell>
                 <TableCell align="right">{oneEvent.intensity}</TableCell>
-                <TableCell align="right">{dateChange(oneEvent.date.toString())}</TableCell>
+                <TableCell align="right">{dateDisplay(oneEvent.date.toString())}</TableCell>
                 <TableCell align="right">
                     { oneEvent.organizer.organizerId === currentOrganizer.organizerId ?
                 <>
