+++ client/src/views/Main.tsx
@@ -12,6 +12,7 @@ import Paper from '@mui/material/Paper';
 
 import axios from 'axios';
 import EventModelForView from '../models/EventModelForView';
+import dateDisplay from '../utils/dates/DateDisplay'
 import { Link } from 'react-router-dom';
 
 const tableStyle = {
@@ -29,26 +30,7 @@ const imageStyle ={
     left: '0',
     zIndex: '-1'
 }
-function dateChange(date: string){
 
-    date = date.slice(0,19);
-    
-    const MM: Array<string> = ["January", "February","March","April","May","June","July","August","September","October","November", "December"]
-    const year: string = date.slice(0, 4);
-    const  month: string = date.slice(5, 7);
-    const day: string = date.slice(8, 10);
-
-    let hour: string = String(parseInt(date.slice(11,13)));
-    if (parseInt(hour) < 10){ hour = "0"+hour }
-
-    let min: string = String(parseInt(date.slice(14, 17)));
-    if (parseInt(min) < 10){ min = "0"+min }
-
-    const a_p: string = parseInt(hour) < 12 ? 'AM' : 'PM'
-    
-    return ''+MM[parseInt(month)-1]+' '+day+', '+year+' '+hour+':'+min+' '+a_p;
-    
-    }
 const Main = () => {
     const [allEvents, setAllEvents] = React.useState<EventModelForView[]>([]);
     const [loaded, setLoaded] = React.useState(false);
@@ -91,7 +73,7 @@ const Main = () => {
                 </TableCell>
                 <TableCell align="right">{oneEvent.type}</TableCell>
                 <TableCell align="right">{oneEvent.intensity}</TableCell>
-                <TableCell align="right">{dateChange(oneEvent.date.toString())}</TableCell>
+                <TableCell align="right">{dateDisplay(oneEvent.date.toString())}</TableCell>
             </TableRow>
             
             ))}
